feat: enforce float64 throughout scientific/numerical path; add JAX_ENABLE_X64 setup
Agent-Logs-Url: https://github.com/bibymaths/phoscrosstalk/sessions/c337c70f-53ab-4e11-8e55-5ab37d8ba32d

+++ phoscrosstalk/steadystate.py
@@ -344,16 +344,16 @@ def _run_steadystate_solve(
     )
 
     args = (
-        jnp.asarray(theta_opt, dtype=jnp.float32),
-        jnp.asarray(Cg, dtype=jnp.float32),
-        jnp.asarray(Cl, dtype=jnp.float32),
+        jnp.asarray(theta_opt, dtype=jnp.float64),
+        jnp.asarray(Cg, dtype=jnp.float64),
+        jnp.asarray(Cl, dtype=jnp.float64),
         jnp.asarray(site_prot_idx, dtype=jnp.int32),
-        jnp.asarray(K_site_kin, dtype=jnp.float32),
-        jnp.asarray(R, dtype=jnp.float32),
-        jnp.asarray(L_alpha, dtype=jnp.float32),
+        jnp.asarray(K_site_kin, dtype=jnp.float64),
+        jnp.asarray(R, dtype=jnp.float64),
+        jnp.asarray(L_alpha, dtype=jnp.float64),
         jnp.asarray(kin_to_prot_idx, dtype=jnp.int32),
-        jnp.asarray(receptor_mask_prot, dtype=jnp.float32),
-        jnp.asarray(receptor_mask_kin, dtype=jnp.float32),
+        jnp.asarray(receptor_mask_prot, dtype=jnp.float64),
+        jnp.asarray(receptor_mask_kin, dtype=jnp.float64),
         jnp.asarray(prev_site_idx, dtype=jnp.int32),
     )
 
@@ -362,8 +362,8 @@ def _run_steadystate_solve(
         k_act_fn=k_act_fn, s_prod_fn=s_prod_fn, rna_relax=rna_relax,
     )
     term = diffrax.ODETerm(rhs_fn)
-    t_eval = jnp.asarray(t_long, dtype=jnp.float32)
-    y0_jax = jnp.asarray(x0, dtype=jnp.float32)
+    t_eval = jnp.asarray(t_long, dtype=jnp.float64)
+    y0_jax = jnp.asarray(x0, dtype=jnp.float64)
     saveat = diffrax.SaveAt(ts=t_eval)
     stepsize_ctrl = make_stepsize_controller(rtol=rtol, atol=atol)
     solver = make_diffrax_solver("tsit5")
